Move values to another module
